# apps/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "manhwa-backend",
    "description": "Ikiru Bot manhwa scraper API. Backend runs fully on local VPS Postgres. Use Bearer token for protected endpoints. RSS feeds support dual-read params: format (manhwa/manhua/manga) + type (legacy alias), country (KR/CN/JP) + origin (legacy alias).",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/whitelist": {
      "get": {
        "summary": "Whitelist Get",
        "operationId": "whitelist_get_api_v1_whitelist_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Page number (1-based)."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "minimum": 1,
              "maximum": 10000
            },
            "description": "Rows per page (1..10000)."
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ikiru",
                "shinigami",
                "voratoon"
              ]
            },
            "description": "Filter by source."
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive title search."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Keyset cursor (created_at ISO) for pagination \u2014 preferred over page for large tables."
          },
          {
            "name": "merge",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Merge cross-source rows into one (groups sources into sources[] array)."
          }
        ]
      },
      "post": {
        "summary": "Whitelist Post",
        "operationId": "whitelist_post_api_v1_whitelist_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Whitelist Delete",
        "operationId": "whitelist_delete_api_v1_whitelist_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "summary": "Whitelist Patch",
        "operationId": "whitelist_patch_api_v1_whitelist_patch",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/rss": {
      "get": {
        "summary": "Rss",
        "operationId": "rss_api_v1_rss_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Page number (1-based)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            },
            "description": "Results per page (1..100)."
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            },
            "description": "Group chapters by (titleKey, source)."
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "manhwa",
                "manhua",
                "manga"
              ]
            },
            "description": "Filter by format (manhwa/manhua/manga). Dual-read: accepts 'type' as legacy alias."
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "manhwa",
                "manhua",
                "manga"
              ]
            },
            "description": "Legacy alias for format. Prefer 'format'."
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "KR",
                "CN",
                "JP"
              ]
            },
            "description": "Filter by country of origin. Dual-read: accepts 'origin' as legacy alias."
          },
          {
            "name": "origin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "KR",
                "CN",
                "JP"
              ]
            },
            "description": "Legacy alias for country. Prefer 'country'."
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ikiru",
                "shinigami",
                "voratoon"
              ]
            },
            "description": "Filter by source."
          },
          {
            "name": "exclude",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exclude country codes (comma-separated, e.g. JP)."
          },
          {
            "name": "exclude_origin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Legacy alias for exclude."
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": "Search title (case-insensitive ILIKE)."
          },
          {
            "name": "genres",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by genre (comma-separated)."
          },
          {
            "name": "min_rating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 10
            },
            "description": "Minimum rating filter."
          },
          {
            "name": "max_rating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 10
            },
            "description": "Maximum rating filter."
          },
          {
            "name": "subscribed_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Only show whitelisted series."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "newest",
                "rating",
                "popular"
              ],
              "default": "newest"
            },
            "description": "Sort order."
          },
          {
            "name": "whitelist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Only whitelisted (same as subscribed_only)."
          },
          {
            "name": "exclude_notified",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Exclude chapters already dispatched."
          },
          {
            "name": "unread_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Exclude already-sent chapters."
          }
        ]
      }
    },
    "/api/v1/reader/rss": {
      "get": {
        "summary": "Rss Reader",
        "description": "Alias for FE compatibility \u2014 /api/v1/reader/rss \u2192 /api/v1/rss.",
        "operationId": "rss_reader_api_v1_reader_rss_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Page number (1-based)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            },
            "description": "Results per page (1..100)."
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            },
            "description": "Group chapters by (titleKey, source)."
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "manhwa",
                "manhua",
                "manga"
              ]
            },
            "description": "Filter by format (manhwa/manhua/manga). Dual-read: accepts 'type' as legacy alias."
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "manhwa",
                "manhua",
                "manga"
              ]
            },
            "description": "Legacy alias for format. Prefer 'format'."
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "KR",
                "CN",
                "JP"
              ]
            },
            "description": "Filter by country of origin. Dual-read: accepts 'origin' as legacy alias."
          },
          {
            "name": "origin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "KR",
                "CN",
                "JP"
              ]
            },
            "description": "Legacy alias for country. Prefer 'country'."
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ikiru",
                "shinigami",
                "voratoon"
              ]
            },
            "description": "Filter by source."
          },
          {
            "name": "exclude",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exclude country codes (comma-separated, e.g. JP)."
          },
          {
            "name": "exclude_origin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Legacy alias for exclude."
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": "Search title (case-insensitive ILIKE)."
          },
          {
            "name": "genres",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by genre (comma-separated)."
          },
          {
            "name": "min_rating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 10
            },
            "description": "Minimum rating filter."
          },
          {
            "name": "max_rating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 10
            },
            "description": "Maximum rating filter."
          },
          {
            "name": "subscribed_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Only show whitelisted series."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "newest",
                "rating",
                "popular"
              ],
              "default": "newest"
            },
            "description": "Sort order."
          },
          {
            "name": "whitelist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Only whitelisted (same as subscribed_only)."
          },
          {
            "name": "exclude_notified",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Exclude chapters already dispatched."
          },
          {
            "name": "unread_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Exclude already-sent chapters."
          }
        ]
      }
    },
    "/api/v1/rss/new": {
      "get": {
        "summary": "Rss New",
        "description": "Lightweight new-items counter.",
        "operationId": "rss_new_api_v1_rss_new_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "https://scanner.aldifhr.fun",
      "description": "Production (VPS)"
    }
  ]
}